+++ fbw-a32nx/src/systems/failures/src/failures-orchestrator.ts
@@ -2,7 +2,11 @@
 //
 // SPDX-License-Identifier: GPL-3.0
 
+import { EventBus } from '@microsoft/msfs-sdk';
+
 import { AtaChapterNumber } from '@flybywiresim/fbw-sdk';
+
+import { RandomFailureGen } from './RandomFailureGen';
 import { QueuedSimVarWriter, SimVarReaderWriter } from './communication';
 import { getActivateFailureSimVarName, getDeactivateFailureSimVarName } from './sim-vars';
 
@@ -20,6 +24,8 @@ export interface Failure {
 export class FailuresOrchestrator {
     private failures: Failure[] = [];
 
+    private randomFailureGe: RandomFailureGen;
+
     private activeFailures = new Set<number>();
 
     private changingFailures = new Set<number>();
@@ -28,7 +34,8 @@ export class FailuresOrchestrator {
 
     private deactivateFailureQueue: QueuedSimVarWriter;
 
-    constructor(simVarPrefix: string, failures: [AtaChapterNumber, number, string][]) {
+    constructor(bus: EventBus, simVarPrefix: string, failures: [AtaChapterNumber, number, string][]) {
+        this.randomFailureGe = new RandomFailureGen(bus);
         this.activateFailureQueue = new QueuedSimVarWriter(new SimVarReaderWriter(getActivateFailureSimVarName(simVarPrefix)));
         this.deactivateFailureQueue = new QueuedSimVarWriter(new SimVarReaderWriter(getDeactivateFailureSimVarName(simVarPrefix)));
         failures.forEach((failure) => {
@@ -41,6 +48,7 @@ export class FailuresOrchestrator {
     }
 
     update() {
+        this.randomFailureGe.update(this);
         this.activateFailureQueue.update();
         this.deactivateFailureQueue.update();
     }