+++ fbw-a32nx/src/systems/instruments/src/EFB/UtilComponents/Modals/Modals.tsx
@@ -2,7 +2,7 @@
 import React, { createContext, FC, useContext, useState } from 'react';
 import { t } from '../../translation';
 
-interface ModalContextInterface{
+export interface ModalContextInterface{
     showModal: (modal: JSX.Element) => void;
     modal?: JSX.Element;
     popModal: () => void;
@@ -70,19 +70,19 @@ export const PromptModal: FC<PromptModalProps> = ({
 
     return (
 
-        <div className="p-8 w-5/12 bg-theme-body rounded-xl border-2 border-theme-accent">
+        <div className="p-8 w-5/12 rounded-xl border-2 bg-theme-body border-theme-accent">
             <h1 className="font-bold">{title}</h1>
             <p className="mt-4">{bodyText}</p>
 
             <div className="flex flex-row mt-8 space-x-4">
                 <div
-                    className="flex justify-center items-center py-2 px-8 w-full text-center text-theme-text hover:text-theme-highlight bg-theme-accent hover:bg-theme-body rounded-md border-2 border-theme-accent hover:border-theme-highlight transition duration-100"
+                    className="flex justify-center items-center py-2 px-8 w-full text-center rounded-md border-2 transition duration-100 text-theme-text hover:text-theme-highlight bg-theme-accent hover:bg-theme-body border-theme-accent hover:border-theme-highlight"
                     onClick={handleCancel}
                 >
                     {cancelText ?? t('Modals.Cancel')}
                 </div>
                 <div
-                    className="flex justify-center items-center py-2 px-8 w-full text-center text-theme-body hover:text-theme-highlight bg-theme-highlight hover:bg-theme-body rounded-md border-2 border-theme-highlight transition duration-100"
+                    className="flex justify-center items-center py-2 px-8 w-full text-center rounded-md border-2 transition duration-100 text-theme-body hover:text-theme-highlight bg-theme-highlight hover:bg-theme-body border-theme-highlight"
                     onClick={handleConfirm}
                 >
                     {confirmText ?? t('Modals.Confirm')}
@@ -106,11 +106,11 @@ export const AlertModal: FC<AlertModalProps> = ({
     };
 
     return (
-        <div className="p-8 w-5/12 bg-theme-body rounded-xl border-2 border-theme-accent">
+        <div className="p-8 w-5/12 rounded-xl border-2 bg-theme-body border-theme-accent">
             <h1 className="font-bold">{title}</h1>
             <p className="mt-4">{bodyText}</p>
             <div
-                className="flex justify-center items-center py-2 px-8 mt-8 w-full text-center text-theme-body hover:text-theme-highlight bg-theme-highlight hover:bg-theme-body rounded-md border-2 border-theme-highlight transition duration-100"
+                className="flex justify-center items-center py-2 px-8 mt-8 w-full text-center rounded-md border-2 transition duration-100 text-theme-body hover:text-theme-highlight bg-theme-highlight hover:bg-theme-body border-theme-highlight"
                 onClick={handleAcknowledge}
             >
                 {acknowledgeText ?? t('Modals.Okay')}
@@ -124,7 +124,7 @@ export const ModalContainer = () => {
 
     return (
         <div className={`fixed inset-0 z-50 transition duration-200 ${modal ? 'opacity-100' : 'opacity-0 pointer-events-none'}`}>
-            <div className="absolute inset-0 bg-theme-body opacity-75" />
+            <div className="absolute inset-0 opacity-75 bg-theme-body" />
             <div className="flex absolute inset-0 flex-col justify-center items-center">
                 {modal}
             </div>